♻️ replace trackViews and trackClickActions with their internal API ports, remove the old startRum view/action glue (PoC phase 3 consolidation)
