+++ packages/browser-rum-core/src/domain/contexts/featureFlagContext.spec.ts
@@ -4,10 +4,10 @@ import { createHook } from '@datadog/js-core/assembly'
 import type { Clock } from '@datadog/browser-core/test'
 import { mockClock } from '@datadog/browser-core/test'
 import { LifeCycle, LifeCycleEventType } from '../lifeCycle'
-import type { ViewCreatedEvent } from '../view/trackViews'
 import type { RumConfiguration } from '../configuration'
 import { RumEventType } from '../../rawRumEvent.types'
 import type { AssembleHook, AssembleHookParams } from '../hooks'
+import type { ViewCreatedEvent } from './viewHistory'
 import type { FeatureFlagContexts } from './featureFlagContext'
 import { startFeatureFlagContexts } from './featureFlagContext'
