+++ packages/browser-rum-core/src/domain/contexts/internalContext.spec.ts
@@ -1,7 +1,7 @@
 import { type RelativeTime } from '@datadog/js-core/time'
 import { noop, type SessionManager } from '@datadog/browser-core'
 import { buildLocation, createSessionManagerMock } from '@datadog/browser-core/test'
-import type { ActionContexts } from '../action/actionCollection'
+import type { ActionContexts } from './internalContext'
 import { startInternalContext } from './internalContext'
 import type { ViewHistory } from './viewHistory'
 import type { UrlContexts } from './urlContexts'
