--- packages/browser-rum-core/src/domain/view/trackViewEventCounts.ts
@@ -1,15 +1,0 @@
-import type { LifeCycle } from '../lifeCycle'
-import { trackEventCounts } from '../trackEventCounts'
-
-export function trackViewEventCounts(lifeCycle: LifeCycle, viewId: string, onChange: () => void) {
-  const { stop, eventCounts } = trackEventCounts({
-    lifeCycle,
-    isChildEvent: (event) => event.view.id === viewId,
-    onChange,
-  })
-
-  return {
-    stop,
-    eventCounts,
-  }
-}
